+++ spec/unit/recipes/default_spec.rb
@@ -2,34 +2,50 @@
 # Cookbook:: chrony
 # Spec:: default
 #
-# Copyright:: 2018-2019, The Authors, All Rights Reserved.
+# Author: Dang H. Nguyen <dang.nguyen@disney.com>
+# Copyright:: 2020, The Walt Disney Company, All Rights Reserved
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
 require 'spec_helper'
 
 describe 'chrony::default' do
-  context 'When all attributes are default, on Ubuntu 16.04' do
-    let(:chef_run) do
-      # for a complete list of available platforms and versions see:
-      # https://github.com/customink/fauxhai/blob/master/PLATFORMS.md
-      runner = ChefSpec::ServerRunner.new(platform: 'ubuntu', version: '16.04')
-      runner.converge(described_recipe)
+  context 'When all attributes are default' do
+    context 'on Ubuntu 18.04' do
+      let(:chef_run) do
+        ChefSpec::ServerRunner.new(platform: 'ubuntu', version: '18.04') do
+          allow_any_instance_of(Chef::Recipe).to receive(:systemd?).and_return(false)
+          allow_any_instance_of(Chef::Resource).to receive(:systemd?).and_return(false)
+        end.converge(described_recipe)
+      end
+
+      it 'included the client recipe' do
+        expect(chef_run).to include_recipe('chrony::client')
+      end
     end
 
-    it 'converges successfully' do
-      expect { chef_run }.to_not raise_error
-    end
-  end
+    context 'on CentOS 7' do
+      let(:chef_run) do
+        ChefSpec::ServerRunner.new(platform: 'centos', version: '7') do
+          allow_any_instance_of(Chef::Recipe).to receive(:systemd?).and_return(true)
+          allow_any_instance_of(Chef::Resource).to receive(:systemd?).and_return(true)
+        end.converge(described_recipe)
+      end
 
-  context 'When all attributes are default, on CentOS 7.4.1708' do
-    let(:chef_run) do
-      # for a complete list of available platforms and versions see:
-      # https://github.com/customink/fauxhai/blob/master/PLATFORMS.md
-      runner = ChefSpec::ServerRunner.new(platform: 'centos', version: '7.4.1708')
-      runner.converge(described_recipe)
-    end
-
-    it 'converges successfully' do
-      expect { chef_run }.to_not raise_error
+      it 'included the client recipe' do
+        expect(chef_run).to include_recipe('chrony::client')
+      end
     end
   end
 end